+++ csrc/apis/c/detector.h
@@ -40,6 +40,10 @@ typedef struct mm_detect_t {
 MMDEPLOY_API int mmdeploy_detector_create(mm_model_t model, const char* device_name, int device_id,
                                           mm_handle_t* handle);
 
+/**
+ * @brief Same as \ref mmdeploy_detector_create, but allows to control execution context of tasks
+ * via exec_info
+ */
 MMDEPLOY_API int mmdeploy_detector_create_v2(mm_model_t model, const char* device_name,
                                              int device_id, mmdeploy_exec_info_t exec_info,
                                              mm_handle_t* handle);
@@ -55,6 +59,12 @@ MMDEPLOY_API int mmdeploy_detector_create_v2(mm_model_t model, const char* devic
 MMDEPLOY_API int mmdeploy_detector_create_by_path(const char* model_path, const char* device_name,
                                                   int device_id, mm_handle_t* handle);
 
+/**
+ * @brief Pack detector inputs into mmdeploy_value_t
+ * @param[in] mats a batch of images
+ * @param[in] mat_count number of images in the batch
+ * @return the created value
+ */
 MMDEPLOY_API mmdeploy_value_t mmdeploy_detector_create_input(const mm_mat_t* mats, int mat_count);
 
 /**
@@ -71,13 +81,40 @@ MMDEPLOY_API mmdeploy_value_t mmdeploy_detector_create_input(const mm_mat_t* mat
  */
 MMDEPLOY_API int mmdeploy_detector_apply(mm_handle_t handle, const mm_mat_t* mats, int mat_count,
                                          mm_detect_t** results, int** result_count);
-
+/**
+ * @brief Same as \ref mmdeploy_detector_apply, but input and output are packed in \ref
+ * mmdeploy_value_t.
+ */
 MMDEPLOY_API int mmdeploy_detector_apply_v2(mm_handle_t handle, mmdeploy_value_t input,
                                             mmdeploy_value_t* output);
 
+/**
+ * @brief Apply detector asynchronously
+ * @param[in] handle handle to the detector
+ * @param[in] input input sender
+ * @return output sender
+ */
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_detector_apply_async(mm_handle_t handle,
                                                              mmdeploy_sender_t input);
 
+// TODO: for discussion only, not implemented
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_detector_apply_async_v2(mm_handle_t handle,
+                                                                mmdeploy_sender_t* input);
+
+// TODO: for discussion only, not implemented
+MMDEPLOY_API int mmdeploy_detector_apply_async_v3(mm_handle_t handle, mmdeploy_sender_t* input,
+                                                  mmdeploy_sender_t* output);
+
+/**
+ * @brief Unpack detector output from a mmdeploy_value_t
+ * @param[in] output output sender returned by applying a detector
+ * @param[out] results a linear buffer to save detection results of each image. It must be released
+ * by \ref mmdeploy_detector_release_result
+ * @param[out] result_count a linear buffer with length number of input images to save the number of
+ * detection results of each image. Must be released by \ref
+ * mmdeploy_detector_release_result
+ * @return status of the operation
+ */
 MMDEPLOY_API int mmdeploy_detector_get_result(mmdeploy_value_t output, mm_detect_t** results,
                                               int** result_count);
