+++ csrc/apis/c/executor.h
@@ -13,6 +13,10 @@ extern "C" {
 
 typedef mmdeploy_value_t (*mmdeploy_then_fn_t)(mmdeploy_value_t, void*);
 
+typedef mmdeploy_value_t (*mmdeploy_then_fn_v2_t)(mmdeploy_value_t*, void*);
+
+typedef int (*mmdeploy_then_fn_v3_t)(mmdeploy_value_t* input, mmdeploy_value_t* output, void*);
+
 struct mmdeploy_sender;
 struct mmdeploy_scheduler;
 
@@ -32,13 +36,18 @@ struct mmdeploy_exec_info {
 ///////////////////////////////////////////////////////////////////////////////
 // Scheduler
 ///////////////////////////////////////////////////////////////////////////////
-MMDEPLOY_API mmdeploy_scheduler_t mmdeploy_executor_inlined();
+MMDEPLOY_API mmdeploy_scheduler_t mmdeploy_executor_inline();
 
 MMDEPLOY_API mmdeploy_scheduler_t mmdeploy_executor_system_pool();
 
+/**
+ * Create a thread pool with the given number of worker threads
+ * @param[in] num_threads
+ * @return the handle to the created thread pool
+ */
 MMDEPLOY_API mmdeploy_scheduler_t mmdeploy_executor_create_thread_pool(int num_threads);
 
-MMDEPLOY_API mmdeploy_scheduler_t mmdeploy_executor_create_single_thread();
+MMDEPLOY_API mmdeploy_scheduler_t mmdeploy_executor_create_thread();
 
 MMDEPLOY_API mmdeploy_scheduler_t mmdeploy_executor_dynamic_batch(mmdeploy_scheduler_t scheduler,
                                                                   int max_batch_size, int timeout);
@@ -48,8 +57,25 @@ MMDEPLOY_API int mmdeploy_scheduler_destroy(mmdeploy_scheduler_t scheduler);
 ///////////////////////////////////////////////////////////////////////////////
 // Sender factories
 ///////////////////////////////////////////////////////////////////////////////
+
+/**
+ * @brief Create a sender that sends the provided value
+ * @param[in] value
+ * @return created sender
+ */
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_just(mmdeploy_value_t value);
 
+// TODO: for discussion only, not implemented
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_just_v2(mmdeploy_value_t* value);
+
+// TODO: for discussion only, not implemented
+MMDEPLOY_API int mmdeploy_executor_just_v3(mmdeploy_value_t* value, mmdeploy_sender_t* sender);
+
+/**
+ * @brief
+ * @param[in] scheduler
+ * @return the sender created
+ */
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_schedule(mmdeploy_scheduler_t scheduler);
 
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_transfer_just(mmdeploy_scheduler_t scheduler,
@@ -58,6 +84,13 @@ MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_transfer_just(mmdeploy_schedule
 ///////////////////////////////////////////////////////////////////////////////
 // Sender adapters
 ///////////////////////////////////////////////////////////////////////////////
+
+/**
+ * Transfer the execution to the execution agent of the provided scheduler
+ * @param[in] input
+ * @param[in] scheduler
+ * @return the sender created
+ */
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_transfer(mmdeploy_sender_t input,
                                                           mmdeploy_scheduler_t scheduler);
 
@@ -67,6 +100,14 @@ MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_on(mmdeploy_scheduler_t schedul
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_then(mmdeploy_sender_t input,
                                                       mmdeploy_then_fn_t fn, void* context);
 
+// TODO: for discussion only, not implemented
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_then_v2(mmdeploy_sender_t* input,
+                                                         mmdeploy_then_fn_v2_t fn, void* context);
+
+// TODO: for discussion only, not implemented
+MMDEPLOY_API int mmdeploy_executor_then_v3(mmdeploy_sender_t* input, mmdeploy_then_fn_v3_t,
+                                           void* context, mmdeploy_sender_t* output);
+
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_let_value(mmdeploy_sender_t input,
                                                            mmdeploy_let_value_fn_t fn,
                                                            void* context);
@@ -74,6 +115,14 @@ MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_let_value(mmdeploy_sender_t inp
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_split(mmdeploy_sender_t input);
 
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_when_all(mmdeploy_sender_t* inputs, int32_t n);
+
+// TODO: for discussion only, not implemented
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_when_all_v2(mmdeploy_sender_t* inputs[],
+                                                             int32_t n);
+
+// TODO: for discussion only, not implemented
+MMDEPLOY_API int mmdeploy_executor_when_all_v3(mmdeploy_sender_t** inputs, int32_t n,
+                                               mmdeploy_sender_t* output);
 
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_ensure_started(mmdeploy_sender_t input);
 
@@ -90,9 +139,24 @@ MMDEPLOY_API void mmdeploy_executor_execute(mmdeploy_scheduler_t scheduler, mmde
 ///////////////////////////////////////////////////////////////////////////////
 // Utilities
 ///////////////////////////////////////////////////////////////////////////////
+
+/**
+ * @brief Create a copy of a copyable sender. Only senders created by \ref mmdeploy_executor_split
+ * is copyable for now.
+ * @param[in] input copyable sender,
+ * @return the sender created, or nullptr if the sender is not copyable
+ */
 MMDEPLOY_API mmdeploy_sender_t mmdeploy_sender_copy(mmdeploy_sender_t input);
 
+/**
+ * @brief Destroy a sender, notice that all sender adapters will consume input senders, only unused
+ * senders should be destroyed using this function.
+ * @param[in] input
+ */
 MMDEPLOY_API int mmdeploy_sender_destroy(mmdeploy_sender_t sender);
+
+// TODO: for discussion only, not implemented
+MMDEPLOY_API int mmdeploy_sender_destroy_v2(mmdeploy_sender_t* sender);
 
 #if __cplusplus
 }
